Change int to long during multiplication.

+++ ColorProfiles/ColorProfiles/ProfileConverter.cs
@@ -87,9 +87,9 @@ namespace ColorProfiles
                 Math.Pow((double) RGB[0], targetProfile.Gamma),
                 Math.Pow((double) RGB[1], targetProfile.Gamma),
                 Math.Pow((double) RGB[2], targetProfile.Gamma) });
-            int R = RGB[0] >= 0 ? Math.Min((int)(RGB[0] * 255), 255) : 0;
-            int G = RGB[1] >= 0 ? Math.Min((int)(RGB[1] * 255), 255) : 0;
-            int B = RGB[2] >= 0 ? Math.Min((int)(RGB[2] * 255), 255) : 0;
+            int R = (int)(RGB[0] >= 0 ? Math.Min((long)(RGB[0] * 255), 255) : 0);
+            int G = (int)(RGB[1] >= 0 ? Math.Min((long)(RGB[1] * 255), 255) : 0);
+            int B = (int)(RGB[2] >= 0 ? Math.Min((long)(RGB[2] * 255), 255) : 0);
             var color = Color.FromArgb(R, G, B);
 
             return color;
